+++ packages/control-property-editor/test/unit/slice.test.ts
@@ -637,13 +637,16 @@ describe('main redux slice', () => {
                                 children: [
                                     {
                                         label: 'test label',
+                                        enabled: true,
                                         children: [
                                             {
                                                 label: 'test label 2',
+                                                enabled: true,
                                                 children: []
                                             },
                                             {
                                                 label: 'test label 3',
+                                                enabled: true,
                                                 children: []
                                             }
                                         ]
@@ -677,13 +680,16 @@ describe('main redux slice', () => {
                             kind: 'nested',
                             children: [
                                 {
+                                    enabled: true,
                                     label: 'test label',
                                     children: [
                                         {
+                                            enabled: true,
                                             label: 'test label 2',
                                             children: []
                                         },
                                         {
+                                            enabled: true,
                                             label: 'test label 3',
                                             children: []
                                         }