+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/registry.ts
@@ -17,6 +17,7 @@ import { AddCustomSectionQuickAction } from '../common/op-add-custom-section';
 import { AddTableActionQuickAction } from '../fe-v2/create-table-action';
 import { AddTableCustomColumnQuickAction } from './create-table-custom-column';
 import { AddPageActionQuickAction } from '../common/create-page-action';
+import { EnableTableFilteringQuickAction } from './lr-enable-table-filtering';
 import { ToggleSemanticDateRangeFilterBar } from './lr-enable-semantic-date-range-filter-bar';
 type PageName = 'listReport' | 'objectPage' | 'analyticalListPage';
 
@@ -48,7 +49,8 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                         ToggleSemanticDateRangeFilterBar,
                         ChangeTableColumnsQuickAction,
                         AddTableActionQuickAction,
-                        AddTableCustomColumnQuickAction
+                        AddTableCustomColumnQuickAction,
+                        EnableTableFilteringQuickAction
                     ],
                     view,
                     key: name + index