+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 1.1.1
+
+### Patch Changes
+
+-   dee9930: fix source system labelling
+
+## 1.1.0
+
+### Minor Changes
+
+-   849ef1e: updates options and conditions for prompts
+
 ## 1.0.26
 
 ### Patch Changes